build: depfile: Add depfile generator
Basically the old config file generator has become the new depfile
generator, with surprisingly little effort. The command line options
have been changed to whatever they should be.

The new config files are now based upon what is in the dep files. We
still need a proper editor for those depfiles, but this is a nice start.

For this commit, I can't be bothered with making config file editors.
That's what text editors are for. Maybe in the future, if I happen to
have some spare time on my hands I'll add it in. (It would improve the
usability quite extensively)

Signed-off-by: Bart Kuivenhoven <bemkuivenhoven@gmail.com>