build: Config: Add updating of depfile.
In the old case, a new depfile would have to be generated, or the flags
had to be added in manually. Now, the flags can simply be added to the
depfile.

A word of caution is in place here. The flags are added as booleans to
the base set. This can be problematic, if the flags are dependent of
other flags. If this is the case, these flags should be moved manually.

With that behind us, this option should make it easier to edit the
depfiles, as the code scans the build files for flags not yet present in
the depfile.

Signed-off-by: Bart Kuivenhoven <bemkuivenhoven@gmail.com>