depfile: Add autoconf ignore option
This allows flags to be turned of in random and allyes configs, in case
the developer wants to implement a test feature that might intentionally
brake the build, or disallow proper functioning of the product.

In general this is bad practice, but sometimes, when testing safety
features, this might be a way of looking for prefered behaviour (IE hang
on possible security breach).

Signed-off-by: Bart Kuivenhoven <bemkuivenhoven@gmail.com>