build: Add default buildfile flag to depfile
In generating the depfile one has to provide a buildfile. In that
depfile there is now a flag, noting the buildfile used to generate the
depfile.

The flag is propagated into the configuration and the final build. This
means that build can now be used while only providing a buildfile
argument to the depfile generator.

Signed-off-by: Bart Kuivenhoven <bemkuivenhoven@gmail.com>