Build: Phase: Add phase states to build structure
Where before everything happened in one big if, and the developer had to
just figure out where in this huge blob of code the bug would occur, now
there is a system that clearly defines what happens in each state of the
build process.

These phases include:
 * InitialPreconfig,		(--gen-depfile)
 * Preconfig,			(--update-depfile)
 * Configure,			(--configure)
 * Compile and
 * Complete.

These phases each manage a global feature of build and if necessary they
can flow into eachother.

Before, if configure was used without a depfile present, Build would try
and fail to generate one in memory. Now the phase is switched to
generate a depfile, with a condition that once the depfile has been
generated, the configuration state will be restored and rerun.

This should make the code much more understandable to the programmer
than the giant if structure that existed before.

Signed-off-by: Bart Kuivenhoven <bemkuivenhoven@gmail.com>