build: makefile: Add option for creating makefiles
To use this option, use --gen-makefile <makefile name>

This will go through the modules in similar fashion to the normal
command issuer, but instead of generating commands and sending them off
to the command runner, it generates target and makemodule objects and
generates a makefile at <makefile name> as specified.

The generated makefile won't have the ability to clean or configure.

Clean options might be added in later, but the configure option will
most likely remain optional to build itself.

Signed-off-by: Bart Kuivenhoven <bemkuivenhoven@gmail.com>